Back end (#1)
* Add bool Text.completed

We need a mechanism for knowing when a given phrase has been
rated enough times; one easy way of doing this is by adding a new
value that handles that. It reduces the normality of the DB, since
it is redundant information.

* Selecting uncompleted, limited and random texts

Also added more dev data to test those cases.

* Add FKs to Tag, /tag checks FK values.

* Update Text.completed integrity constraint

Note: if we change tagging to scoring, then we should change
the integrity constraints to be something more suitable
(e.g. checking the variance of the scores)

* /tag: check num of tags received