vue-ish way of disabling the button to avoid neutral scores (#21)
* Send button is only enabled after clicking on slider.

Issue #10.
If there are multiple sliders we should consider making it so you have
to click on all of them to enable sending the captcha.
Another option is forbidding any slider to be on 0. Are there any true
neutral sentences? Doing this will force users to think if it's even a
little bit positive or negative.

* vue-ish way of disabling the button to avoid neutral scores