Reorder dockerfile to optimize build time.
For each command in a dockerfile, docker mantains a diff-image cache
that can be used for incremental builds. By reordering the instructions
in the dockerfile we can first get all the python dependencies that are
in the requirements.txt file and then copy the rest of the code.
After this commit, one should expect not a full rebuild by just changing
a line of code.
However, adding a new package will force a full image rebuild.